+++ data/scripts/galaxy/ca_world_eater_manager.lua
@@ -1,6 +1,6 @@
 package.path = package.path .. ";data/scripts/lib/?.lua"
 include("stringutility")
-include("callable")
+include("data/scripts/lib/callable")
 local CosmicAscendancyNews = include("ca_news")
 local CosmicVaultData = include("cosmicvaultdata")
 local EncounterBridge = include("ca_encounter_bridge")